+++ packages/react-router/__tests__/data-memory-router-test.tsx
@@ -295,7 +295,6 @@ describe("createMemoryRouter", () => {
       "<div>
         <div>
           parent data
-          child action
           idle
           <div>
             child data